sessions and cookies added

+++ php/filmoteka/views/film-single.tpl
@@ -7,11 +7,10 @@
     <div class="col-auto">
       <img src="
           <?php 
-              echo HOST . 'data/films/min/' . $film['photo']; 
+              echo HOST . 'data/films/full/' . $film['photo']; 
               if ($film['photo'] == "") { echo "default.jpeg";}
           ?>" 
-          alt="Обложка фильма">
-
+          alt="<?php echo $film['title']?>">
     </div>
     <!--// col -->
 
@@ -20,8 +19,12 @@
       <div class="card__header">
         <h4 class="title-4"><?php echo $film['title']?></h4>
         <div class="buttons">
-          <a href="edit.php?id=<?php echo $film['id']?>" class="button button--edit">Редактировать</a>
-          <a href="index.php?action=delete&id=<?php echo $film['id']?>" class="button button--delete">Удалить</a>
+
+          <?php if ( isAdmin () ) { ?>
+                <a href="edit.php?id=<?php echo $film['id']?>" class="button button--edit">Редактировать</a>
+                <a href="index.php?action=delete&id=<?php echo $film['id']?>" class="button button--delete">Удалить</a>
+          <?php } ?>
+          
         </div>
       </div>
       <div class="badge"><?php echo $film['genre']?></div>
